check principal while deleting/updating

+++ src/main/java/com/beerjournal/breweriana/item/ItemService.java
@@ -3,13 +3,13 @@ package com.beerjournal.breweriana.item;
 import com.beerjournal.breweriana.item.persistence.Item;
 import com.beerjournal.breweriana.item.persistence.ItemRepository;
 import com.beerjournal.breweriana.user.persistence.UserRepository;
+import com.beerjournal.breweriana.utils.SecurityUtils;
 import com.beerjournal.breweriana.utils.ServiceUtils;
 import com.beerjournal.infrastructure.error.BeerJournalException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import static com.beerjournal.infrastructure.error.ErrorInfo.ITEM_NOT_FOUND;
-import static com.beerjournal.infrastructure.error.ErrorInfo.USER_NOT_FOUND;
+import static com.beerjournal.infrastructure.error.ErrorInfo.*;
 
 @Service
 @RequiredArgsConstructor
@@ -17,6 +17,7 @@ class ItemService {
 
     private final ItemRepository itemRepository;
     private final UserRepository userRepository;
+    private final SecurityUtils securityUtils;
 
     ItemDto getItemDetails(String id) {
         Item item = itemRepository.findOneById(ServiceUtils.stringToObjectId(id))
@@ -38,6 +39,10 @@ class ItemService {
         userRepository.findOneById(ServiceUtils.stringToObjectId(ownerId))
                 .orElseThrow(() -> new BeerJournalException(USER_NOT_FOUND));
 
+        if (!securityUtils.checkIfAuthorized(ownerId)){
+            throw new BeerJournalException(COLLECTION_FORBIDDEN_MODIFICATION);
+        }
+
         Item deletedItem = itemRepository.delete(itemId);
         return ItemDto.toDto(deletedItem);
     }
@@ -45,6 +50,10 @@ class ItemService {
     ItemDto updateItem(String ownerId, String itemId, ItemDto itemDto) {
         userRepository.findOneById(ServiceUtils.stringToObjectId(ownerId))
                 .orElseThrow(() -> new BeerJournalException(USER_NOT_FOUND));
+
+        if (!securityUtils.checkIfAuthorized(ownerId)){
+            throw new BeerJournalException(COLLECTION_FORBIDDEN_MODIFICATION);
+        }
 
         Item itemToUpdate = Item.copyWithAssignedId(ServiceUtils.stringToObjectId(itemId),
                 ItemDto.fromDto(itemDto, ownerId));
